"diff of user and amdin routes"

+++ controllers/users.controller.js
@@ -4,8 +4,10 @@ const { generateToken } = require("../utils/jwt.utils");
 
 async function createUser(req, res) {
   try {
-    const { firstName, lastName, email, password, age, phoneNumber } = req.data;
+    const { role, firstName, lastName, email, password, age, phoneNumber } =
+      req.data;
     const newUser = new User({
+      role,
       firstName,
       lastName,
       email,
@@ -29,11 +31,13 @@ async function createUser(req, res) {
 async function getAllUser(req, res) {
   try {
     const allUsers = await User.find();
-    if ((allUsers.length = 0)) {
+
+    if (allUsers.length == 0) {
       return res
         .status(500)
         .json({ message: "Users are not availiable in db" });
     }
+    console.log(allUsers);
     return res
       .status(200)
       .json({ message: "All users retrieved successfully: ", allUsers });
@@ -55,11 +59,30 @@ async function getUserById(req, res) {
   }
 }
 
+async function adminUpdateAnyUser(req, res) {
+  try {
+    const updateData = req.body;
+    const resourceId = req.params.id;
+    const updatedUser = await User.findByIdAndUpdate(resourceId, updateData, {
+      new: true,
+      runValidators: true,
+    });
+
+    if (!updatedUser) {
+      return res.status(404).json({ message: "User not found" });
+    }
+    return res.status(200).json({ message: "User is Updated : ", updatedUser });
+  } catch (error) {
+    return res
+      .status(500)
+      .json({ message: " Internal Server Error!", error: error.message });
+  }
+}
+
 async function updateUserByID(req, res) {
   try {
-  // console.log(req.decode,"last")
     const updateData = req.body;
-    
+
     const resourceId = req.decode.id;
     const updatedUser = await User.findByIdAndUpdate(resourceId, updateData, {
       new: true,
@@ -71,12 +94,18 @@ async function updateUserByID(req, res) {
     }
     return res.status(200).json({ message: "User is Updated : ", updatedUser });
   } catch (error) {
-    return res.status(500).json({ message: " Internal Server Error!", error });
+    return res
+      .status(500)
+      .json({ message: " Internal Server Error!", error: error.message });
   }
 }
 
 async function findUserAndDelete(req, res) {
   try {
+    const userCheck = await User.findById(req.params.id);
+    if (userCheck.role === "admin") {
+      return res.status(404).json({ message: "Admin cannot be removed " });
+    }
     const userFoundById = await User.findByIdAndDelete(req.params.id);
 
     if (!userFoundById) {
@@ -88,12 +117,28 @@ async function findUserAndDelete(req, res) {
       .status(200)
       .json({ message: "user is deleted : ", userFoundById });
   } catch (error) {
-    return res
-      .status(404)
-      .json({ message: "User is not found by this id for delete!", error });
+    return res.status(500).json({ message: "Internal Server error!", error });
   }
 }
 
+async function deleteUser(req, res) {
+  try {
+    const userFoundById = await User.findByIdAndDelete(req.decode.id);
+    // console.log(userFoundById,"hehehe");
+    if (!userFoundById) {
+      return res
+        .status(404)
+        .json({ message: "User is not found by this Id for deletion" });
+    }
+    return res
+      .status(200)
+      .json({ message: "user is deleted : ", userFoundById });
+  } catch (error) {
+    return res
+      .status(500)
+      .json({ message: "Internal server error", error: error.message });
+  }
+}
 async function getUserByEmail(req, res) {
   try {
     const userFoundByEmail = await User.findOne({ email: req.params.email });
@@ -117,8 +162,6 @@ async function verifyUser(req, res) {
     if (!verifiedUser) {
       return res.status(404).json({ message: "User not found" });
     }
-    // console.log(verifiedUser);
-    // console.log(verifiedUser._id.toString());
 
     if (verifiedUser.password !== req.data.password) {
       return res
@@ -127,9 +170,10 @@ async function verifyUser(req, res) {
     }
     const token = generateToken({
       id: verifiedUser._id.toString(),
+      role: verifiedUser.role,
       fullName: verifiedUser.fullName,
     });
-    // console.log(token);
+
     return res.status(200).json({ message: "User is verified", token: token });
   } catch (error) {
     return res.status(404).json({ message: "Internal server error", error });
@@ -137,7 +181,6 @@ async function verifyUser(req, res) {
 }
 
 async function getUserByAge(req, res) {
-  //ehehe
   try {
     if (req.query.age <= 0)
       return res.status(500).json({ message: "Age must be more than 0" });
@@ -162,6 +205,7 @@ async function getUserByAgeLesser(req, res) {
     if (req.query.age <= 0)
       return res.status(404).json({ message: "Age must be more than 0" });
     const userFoundByAge = await User.find({ age: { $lte: req.query.age } });
+
     if (!userFoundByAge)
       return res
         .status(404)
@@ -185,4 +229,6 @@ module.exports = {
   getUserByAge,
   getUserByAgeLesser,
   verifyUser,
+  adminUpdateAnyUser,
+  deleteUser,
 };
